Implemented yotams code into repo, file organization
